Resolved merge issue with moved functions
