+++ tests/copilot_helper/test_command_dns.py
@@ -1,6 +1,4 @@
-import json
 from unittest.mock import ANY
-from unittest.mock import mock_open
 from unittest.mock import patch
 
 import boto3
@@ -27,7 +25,6 @@ from dbt_copilot_helper.commands.dns import create_hosted_zones
 from dbt_copilot_helper.commands.dns import create_required_zones_and_certs
 from dbt_copilot_helper.commands.dns import get_base_domain
 from dbt_copilot_helper.commands.dns import get_certificate_zone_id
-from dbt_copilot_helper.commands.dns import get_load_balancer_domain_and_configuration
 from dbt_copilot_helper.commands.dns import get_required_subdomains
 from dbt_copilot_helper.commands.dns import validate_subdomains
 from tests.copilot_helper.conftest import mock_aws_client
@@ -559,148 +556,6 @@ def test_assign(ensure_cwd_is_repo_root, get_aws_session_or_abort):
         ],
     )
     assert result.output.startswith("There are no clusters matching")
-
-
-@mock_ecs
-def test_get_load_balancer_domain_and_configuration_no_clusters(capfd):
-    with pytest.raises(SystemExit):
-        get_load_balancer_domain_and_configuration(
-            boto3.Session(),
-            HYPHENATED_APPLICATION_NAME,
-            ALPHANUMERIC_ENVIRONMENT_NAME,
-            ALPHANUMERIC_SERVICE_NAME,
-        )
-
-    out, _ = capfd.readouterr()
-
-    assert (
-        out == f"There are no clusters matching {HYPHENATED_APPLICATION_NAME} in this AWS account\n"
-    )
-
-
-@patch("dbt_copilot_helper.commands.dns.get_load_balancer_configuration", return_value="test.com")
-def test_get_load_balancer_domain_and_configuration_no_domain(
-    get_load_balancer_configuration, fakefs, capsys
-):
-    fakefs.create_file(
-        "copilot/testsvc1/manifest.yml",
-        contents="""
-environments:
-  test:
-    http:
-      alias:
-""",
-    )
-    with pytest.raises(SystemExit):
-        get_load_balancer_domain_and_configuration("test", "testapp", "testsvc1", "test")
-    assert (
-        capsys.readouterr().out
-        == "No domains found, please check the ./copilot/testsvc1/manifest.yml file\n"
-    )
-
-    fakefs.create_file(
-        "copilot/testsvc2/manifest.yml",
-        contents="""
-environments:
-  test:
-    http:
-""",
-    )
-    with pytest.raises(SystemExit):
-        get_load_balancer_domain_and_configuration("test", "testapp", "testsvc2", "test")
-    assert (
-        capsys.readouterr().out
-        == "No domains found, please check the ./copilot/testsvc2/manifest.yml file\n"
-    )
-
-    fakefs.create_file(
-        "copilot/testsvc3/manifest.yml",
-        contents="""
-environments:
-  not_test:
-    http:
-        alias: test.com
-""",
-    )
-    with pytest.raises(SystemExit):
-        get_load_balancer_domain_and_configuration("test", "testapp", "testsvc3", "test")
-    assert (
-        capsys.readouterr().out
-        == "Environment test not found, please check the ./copilot/testsvc3/manifest.yml file\n"
-    )
-
-
-@mock_ecs
-def test_get_load_balancer_domain_and_configuration_no_services(capfd):
-    boto3.Session().client("ecs").create_cluster(
-        clusterName=f"{HYPHENATED_APPLICATION_NAME}-{ALPHANUMERIC_ENVIRONMENT_NAME}-{CLUSTER_NAME_SUFFIX}"
-    )
-    with pytest.raises(SystemExit):
-        get_load_balancer_domain_and_configuration(
-            boto3.Session(),
-            HYPHENATED_APPLICATION_NAME,
-            ALPHANUMERIC_SERVICE_NAME,
-            ALPHANUMERIC_ENVIRONMENT_NAME,
-        )
-
-    out, _ = capfd.readouterr()
-
-    assert (
-        out == f"There are no services matching {ALPHANUMERIC_SERVICE_NAME} in this aws account\n"
-    )
-
-
-@mock_elbv2
-@mock_ec2
-@mock_ecs
-def test_get_load_balancer_domain_and_configuration(tmp_path):
-    cluster_name = (
-        f"{HYPHENATED_APPLICATION_NAME}-{ALPHANUMERIC_ENVIRONMENT_NAME}-{CLUSTER_NAME_SUFFIX}"
-    )
-    service_name = f"{HYPHENATED_APPLICATION_NAME}-{ALPHANUMERIC_ENVIRONMENT_NAME}-{ALPHANUMERIC_SERVICE_NAME}-{SERVICE_NAME_SUFFIX}"
-    session = boto3.Session()
-    mocked_vpc_id = session.client("ec2").create_vpc(CidrBlock="10.0.0.0/16")["Vpc"]["VpcId"]
-    mocked_subnet_id = session.client("ec2").create_subnet(
-        VpcId=mocked_vpc_id, CidrBlock="10.0.0.0/16"
-    )["Subnet"]["SubnetId"]
-    mocked_elbv2_client = session.client("elbv2")
-    mocked_load_balancer_arn = mocked_elbv2_client.create_load_balancer(
-        Name="foo", Subnets=[mocked_subnet_id]
-    )["LoadBalancers"][0]["LoadBalancerArn"]
-    target_group = mocked_elbv2_client.create_target_group(
-        Name="foo", Protocol="HTTPS", Port=80, VpcId=mocked_vpc_id
-    )
-    target_group_arn = target_group["TargetGroups"][0]["TargetGroupArn"]
-    mocked_elbv2_client.create_listener(
-        LoadBalancerArn=mocked_load_balancer_arn,
-        DefaultActions=[{"Type": "forward", "TargetGroupArn": target_group_arn}],
-    )
-    mocked_ecs_client = session.client("ecs")
-    mocked_ecs_client.create_cluster(clusterName=cluster_name)
-    mocked_ecs_client.create_service(
-        cluster=cluster_name,
-        serviceName=service_name,
-        loadBalancers=[{"loadBalancerName": "foo", "targetGroupArn": target_group_arn}],
-    )
-    mocked_service_manifest_contents = {
-        "environments": {ALPHANUMERIC_ENVIRONMENT_NAME: {"http": {"alias": "somedomain.tld"}}}
-    }
-    open_mock = mock_open(read_data=json.dumps(mocked_service_manifest_contents))
-
-    with patch("dbt_copilot_helper.commands.dns.open", open_mock):
-        domain_name, load_balancer_configuration = get_load_balancer_domain_and_configuration(
-            boto3.Session(),
-            HYPHENATED_APPLICATION_NAME,
-            ALPHANUMERIC_SERVICE_NAME,
-            ALPHANUMERIC_ENVIRONMENT_NAME,
-        )
-
-    open_mock.assert_called_once_with(f"./copilot/{ALPHANUMERIC_SERVICE_NAME}/manifest.yml", "r")
-    assert domain_name == "somedomain.tld"
-    assert load_balancer_configuration["LoadBalancerArn"] == mocked_load_balancer_arn
-    assert load_balancer_configuration["LoadBalancerName"] == "foo"
-    assert load_balancer_configuration["VpcId"] == mocked_vpc_id
-    assert load_balancer_configuration["AvailabilityZones"][0]["SubnetId"] == mocked_subnet_id
 
 
 @pytest.mark.parametrize(
